refactor: extracted unseen provider to separated file
