+++ packages/notification-center/src/components/notification-center/NotificationCenter.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
 import { IMessage, IMessageAction, ButtonTypeEnum } from '@novu/shared';
-import { NotificationCenterContext } from '../../store/notification-center.context';
-import { AppContent } from './components';
 import { QueryClient, QueryClientProvider } from 'react-query';
+import { AppContent } from './components';
 import { useNovuContext } from '../../hooks';
-import { INovuThemeProvider, NovuThemeProvider } from '../../store/novu-theme-provider.context';
+import { INovuThemeProvider, NovuThemeProvider, NotificationCenterContext } from '../../store';
 import { ColorScheme, ITab, ListItem } from '../../index';
 
 export interface INotificationCenterProps {
