+++ packages/notification-center/src/components/novu-provider/NovuProvider.tsx
@@ -122,26 +122,3 @@ function SocketInitialization({ children }: { children: React.ReactNode }) {
 
   return <SocketContext.Provider value={{ socket }}>{children}</SocketContext.Provider>;
 }
-
-function UnseenProvider({ children }: { children: React.ReactNode }) {
-  const { unseenCount, setUnseenCount } = useUnseenController();
-  const { socket } = useSocket();
-
-  useEffect(() => {
-    if (socket) {
-      socket.on('unseen_count_changed', (onData: { unseenCount: number }) => {
-        if (!isNaN(onData?.unseenCount)) {
-          setUnseenCount(onData.unseenCount);
-        }
-      });
-    }
-
-    return () => {
-      if (socket) {
-        socket.off('unseen_count_changed');
-      }
-    };
-  }, [socket]);
-
-  return <UnseenCountContext.Provider value={{ unseenCount, setUnseenCount }}>{children}</UnseenCountContext.Provider>;
-}
